requirements.txt
httpx>=0.24
Pillow>=10.0
